Renamed and refactored LocalizationManagement to make it easier to use. Now ProfileView uses it too.
