+++ admin-web/src/main/java/hu/schonherz/admin/web/locale/LocaleManagerBean.java
@@ -2,7 +2,6 @@ package hu.schonherz.admin.web.locale;
 
 import java.util.Locale;
 
-import javax.annotation.PostConstruct;
 import javax.faces.bean.ManagedBean;
 import javax.faces.bean.SessionScoped;
 import javax.faces.context.FacesContext;
@@ -14,25 +13,18 @@ import lombok.extern.slf4j.Slf4j;
 @SessionScoped
 public class LocaleManagerBean {
 
-    private Locale locale;
-
-    @PostConstruct
-    public void init() {
-        locale = FacesContext.getCurrentInstance().getExternalContext().getRequestLocale();
-    }
-
     public Locale getLocale() {
-        return locale;
+        return LocalizationManagement.getLocale();
     }
 
     public String getLanguage() {
-        return locale.getLanguage();
+        return LocalizationManagement.getLanguage();
     }
 
-    public void setLanguage(String language) {
-        locale = new Locale(language);
-        LocaleManagement.setLanguage(language);
-        FacesContext.getCurrentInstance().getViewRoot().setLocale(locale);
+    public void setLanguage(final String language) {
+        if (LocalizationManagement.setLanguage(language)) {
+            FacesContext.getCurrentInstance().getViewRoot().setLocale(LocalizationManagement.getLocale());
+        }
     }
 
 }
