+++ admin-web/src/main/java/hu/schonherz/project/admin/web/view/RegistrationView.java
@@ -1,6 +1,6 @@
 package hu.schonherz.project.admin.web.view;
 
-import hu.schonherz.admin.web.locale.LocaleManagement;
+import hu.schonherz.admin.web.locale.LocalizationManagement;
 import java.util.ResourceBundle;
 
 import javax.annotation.PostConstruct;
@@ -16,9 +16,7 @@ import hu.schonherz.project.admin.service.api.service.InvalidUserDataException;
 import hu.schonherz.project.admin.service.api.vo.UserRole;
 import hu.schonherz.project.admin.service.api.vo.UserVo;
 import hu.schonherz.project.admin.web.view.form.RegistrationForm;
-import java.util.Locale;
-import javafx.beans.value.ChangeListener;
-import javafx.beans.value.ObservableValue;
+import javafx.beans.property.ObjectProperty;
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 
@@ -39,7 +37,8 @@ public class RegistrationView {
 
     // Wired to the registration xhtml
     private RegistrationForm form;
-    private ResourceBundle localMessages;
+
+    private ObjectProperty<ResourceBundle> messageProperty;
 
     @EJB
     private UserServiceRemote userServiceRemote;
@@ -47,17 +46,12 @@ public class RegistrationView {
     @PostConstruct
     public void init() {
         form = new RegistrationForm();
-        refreshLocaleMessages(Locale.getDefault());
-
-        ChangeListener<Locale> listener = (ObservableValue<? extends Locale> ov, Locale oldLocale, Locale newLocale) -> {
-            refreshLocaleMessages(newLocale);
-        };
-
-        LocaleManagement.addListener(listener);
+        messageProperty = LocalizationManagement.getMessageProperty();
     }
 
     public void registration() {
         FacesContext context = FacesContext.getCurrentInstance();
+        ResourceBundle localMessages = messageProperty.get();
 
         try {
             UserVo userVo = form.getUserVo();
@@ -85,17 +79,6 @@ public class RegistrationView {
     private void setDefaultValues(final UserVo vo) {
         vo.setActive(true);
         vo.setUserRole(UserRole.AGENT);
-    }
-
-    public void refreshLocaleMessages(Locale newLocale) {
-        try {
-            localMessages = ResourceBundle.getBundle("i18n.localization", newLocale);
-        } catch (Exception e) {
-            String message = "Could not create resource bundle for localization messages!";
-            log.error(message, e);
-            throw new IllegalStateException(message, e);
-        }
-
     }
 
 }
